pyproject.toml
[tool.poetry]
name = "micasa"
version = "0.1.0"
description = ""
authors = ["Pierre-Yves Gillier <github@pygillier.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-talisman = "^1.1.0"
flask-migrate = "^4.0.7"
flask-wtf = "^1.2.1"
flask-security-too = {extras = ["common", "fsqla"], version = "^5.4.3"}
babel = "^2.14.0"


[tool.poetry.group.dev.dependencies]
flask-debugtoolbar = "^0.14.1"
ruff = "^0.3.5"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
